Epitech Experience

+++ Plugins/RoukaVici/Source/RoukaVici/Private/RoukaViciManager.cpp
@@ -4,10 +4,10 @@
 
 #include "Runtime/Core/Public/HAL/PlatformFilemanager.h"
 #include "Runtime/JsonUtilities/Public/JsonObjectConverter.h"
-#include "VibrationSelectionWidget.h"
+#include "RoukaViciWidget.h"
 
 URoukaViciManager *URoukaViciManager::instance = NULL;
-UVibrationSelectionWidget *URoukaViciManager::patternEditor = NULL;
+URoukaViciWidget *URoukaViciManager::UI = NULL;
 
 // Sets default values
 URoukaViciManager::URoukaViciManager()
@@ -24,7 +24,7 @@ URoukaViciManager::URoukaViciManager()
 void URoukaViciManager::BeginPlay()
 {
 	Super::BeginPlay();
-	ULibRoukaVici::loadLib();
+	ULibRoukaVici::LoadLib();
 }
 
 // Called every frame
@@ -36,17 +36,22 @@ void URoukaViciManager::TickComponent(float DeltaTime, ELevelTick TickType, FAct
 void URoukaViciManager::BeginDestroy()
 {
 	Super::BeginDestroy();
-	ULibRoukaVici::callStopLib();
+	//ULibRoukaVici::CallStopLib();
 }
 
-void URoukaViciManager::DisplayPatternEditor()
+void URoukaViciManager::DisplayUI()
 {
-	patternEditor->SetVisibility(ESlateVisibility::Visible);
+	UI->SetVisibility(ESlateVisibility::Visible);
 }
 
-void URoukaViciManager::HidePatternEditor()
+void URoukaViciManager::HideUI()
 {
-	patternEditor->SetVisibility(ESlateVisibility::Hidden);
+	UI->SetVisibility(ESlateVisibility::Hidden);
+}
+
+void URoukaViciManager::ToggleUI()
+{
+	UI->toggleUIDelegate.Broadcast();
 }
 
 void URoukaViciManager::SetVibrationPattern(int ID)
@@ -56,6 +61,7 @@ void URoukaViciManager::SetVibrationPattern(int ID)
 
 void URoukaViciManager::SavePattern(const FmPattern &pattern, int editedPattern)
 {
+	// Get the right folder path
 	FString folderPath;
 	if (GetWorld()->WorldType == EWorldType::PIE)
 	{
@@ -63,17 +69,52 @@ void URoukaViciManager::SavePattern(const FmPattern &pattern, int editedPattern)
 		FPaths::NormalizeDirectoryName(folderPath);
 	}
 	folderPath += "/Vibration Patterns/";
-	FString oldPattern = folderPath + patterns[editedPattern].name + ".json";
 	FString newPattern = folderPath + pattern.name + ".json";
-
-	if (oldPattern != newPattern)
-		FPlatformFileManager::Get().GetPlatformFile().DeleteFile(*oldPattern);
-
 	FString OutputString;
-	FJsonObjectConverter::UStructToJsonObjectString(pattern, OutputString, 0, 0, 0, NULL, true);
+
+	if (editedPattern < 0)
+	{
+		// New pattern to save
+		patterns.Add(pattern);
+		FJsonObjectConverter::UStructToJsonObjectString(pattern, OutputString, 0, 0, 0, NULL, true);
+	}
+	else
+	{
+		// Edited Pattern has to replace the previous one
+		FString oldPattern = folderPath + patterns[editedPattern].name + ".json";
+		if (oldPattern != newPattern)
+			FPlatformFileManager::Get().GetPlatformFile().DeleteFile(*oldPattern);
+
+		FJsonObjectConverter::UStructToJsonObjectString(pattern, OutputString, 0, 0, 0, NULL, true);
+		patterns[editedPattern] = pattern;
+	}
+
+	// Write in the Json file
 	FFileHelper::SaveStringToFile(OutputString, *newPattern);
 
-	patterns[editedPattern] = pattern;
-
+	// Tell the UI that it has to update its content
 	updateUIDelegate.Broadcast();
+}
+
+void URoukaViciManager::DeletePattern(int patternToDelete)
+{
+	// Get the right folder path
+	FString folderPath;
+	if (GetWorld()->WorldType == EWorldType::PIE)
+	{
+		folderPath = FPaths::ProjectDir();
+		FPaths::NormalizeDirectoryName(folderPath);
+	}
+	folderPath += "/Vibration Patterns/";
+
+	// Make sure it's in range
+	if (patternToDelete >= patterns.Num() || patternToDelete < 0)
+		return;
+
+	// Delete the Json file
+	FString pattern = folderPath + patterns[patternToDelete].name + ".json";
+	FPlatformFileManager::Get().GetPlatformFile().DeleteFile(*pattern);
+
+	// Remove the pattern from the manager
+	patterns.RemoveAt(patternToDelete);
 }
